OLH-4066: improvements
