+++ vitest.config.ts
@@ -2,8 +2,6 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    globals: true,
-    clearMocks: true,
     setupFiles: ["src/tests/setup.ts"],
     coverage: {
       provider: "v8",
